third_party/WebKit/LayoutTests/external/wpt/html/browsers/browsing-the-web/navigating-across-documents/013-expected.txt

This is a testharness.js-based test.
FAIL Link with onclick navigation to javascript url with delayed document.write and href navigation  assert_array_equals: property 0, expected "after script" but got "javascript"
Harness: the test ran to completion.
